fix(ci): python-coverage-shards — split oversized first-party files under the maintainability line limit
Extract pre-push validation fix-pass helpers into a new
pre_push_validation_fix_pass sub-module (re-exported by the parent to
preserve the public surface and test monkeypatch semantics), and split
the two oversized test modules (test_pr_monitor_pre_push_validation and
test_pr_monitor_runner_coverage_edges_part_005) into new sibling parts so
all first-party files stay under MAX_FIRST_PARTY_FILE_LINES (1500).
